[EI-241] [api] refactor: 파일 경로 수정
